requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
pandas
psycopg2-binary
python-dotenv
